Changing file size detection to support symlinks
Matlab's dir command returns the number of bytes of the symlink itself
rather than the link target. If a raw data file is processed by
Kilosort2 via a symlink, get_file_size returns the correct file size by
using the stat command. On Windows, dir continues to be used (this could
be fixed if people are using MKLink).